attach: verify the jq download against a sha256sum
The downloaded binary was being trusted as-is, which means stedolan may be
capable of adding binaries that run with root privaleges to arbitrary newly
provision packet hosts without anyone really being any the wiser.

This patch assumes that 8 space indentation is ok, which is used in some
places in the file, tabs are elsewhere.

This patch is submitted under the assumption of either an MIT or a BSD
license and should be treated as such. The repository needs to be marked with
a license and a CLA.